isolate profiler support to v1
